dom/imptests/webapps/DOMCore/tests/submissions/Ms2ger/test_CharacterData-deleteData.html

<!DOCTYPE html>
<title>CharacterData.deleteData</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
function testNode(node) {
  test(function() {
    assert_throws("INDEX_SIZE_ERR", function() { node.deleteData(5, 10) })
    assert_throws("INDEX_SIZE_ERR", function() { node.deleteData(5, 0) })
    node.deleteData(2, 10)
    assert_equals(node.data, "te")
    node.data = "test"
    node.deleteData(1, 1)
    assert_equals(node.data, "tst")
  })
}
test(function() {
  testNode(document.createTextNode("test"))
  testNode(document.createComment("test"))
})
</script>
